requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
concurrent-log-handler==0.9.9
elasticsearch==7.8.0
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.1.1
Jinja2==2.11.2
kafka-python==1.4.6
MarkupSafe==1.1.1
more-itertools==8.4.0
nb-log==3.4
packaging==20.4
pdfminer3k==1.3.4
pika==1.1.0
pluggy==0.13.1
ply==3.11
py==1.9.0
pymongo==3.5.1
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-repeat==0.7.0
pywin32==228
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
toml==0.10.2
tomorrow3==1.1.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
